release(1.2.6): update mf6 version
Rename autotests with more descriptive names
