+++ .github/intel-scripts/ifortvars_linux.sh
@@ -14,5 +14,5 @@ ifort --version
 icc --version
 
 # run pytest
-pytest -v -n=auto --dist=loadfile -m=regression --durations=0 --cov=pymake --cov-report=xml autotest/
+pytest -v -n=auto --dist=loadfile -m="base or regression" --durations=0 --cov=pymake --cov-report=xml autotest/
