+++ .github/workflows/pymake-requests.yml
@@ -32,7 +32,7 @@ jobs:
 
       - name: Run pytest
         run: |
-          pytest -v --dist=loadfile -n=auto -m requests --durations=0 --cov=pymake --cov-report=xml autotest/
+          pytest -v -n=auto -m requests --durations=0 --cov=pymake --cov-report=xml autotest/
 
       - name: Print coverage report before upload
         run: |
