+++ autotest/test_gridgen.py
@@ -2,42 +2,46 @@ import os
 import shutil
 import subprocess
 import sys
+import pathlib as pl
 
 import pytest
 
 import pymake
 
+# use the line below to set fortran compiler using environmental variables
+# if sys.platform.lower() == "win32":
+#     os.environ["CC"] = "icl"
+# else:
+#     os.environ["CC"] = "icc"
+
 # define program data
 target = "gridgen"
+if sys.platform.lower() == "win32":
+    target += ".exe"
 
 # get program dictionary
 prog_dict = pymake.usgs_program_data.get_target(target)
 
 # set up paths
-dstpth = os.path.join(f"temp_{os.path.basename(__file__).replace('.py', '')}")
-if not os.path.exists(dstpth):
-    os.makedirs(dstpth, exist_ok=True)
+dstpth = pl.Path(f"temp_{os.path.basename(__file__).replace('.py', '')}")
+dstpth.mkdir(parents=True, exist_ok=True)
 
 ver = prog_dict.version
-pth = os.path.join(dstpth, prog_dict.dirname)
-expth = os.path.join(pth, "examples", "biscayne")
+pth = dstpth / prog_dict.dirname
+expth = pth / "examples/biscayne"
+exe_name = dstpth / target
 
 pm = pymake.Pymake(verbose=True)
 pm.target = target
-pm.appdir = os.path.join(dstpth, "bin")
+pm.appdir = str(dstpth)
 env_var = os.environ.get("CC")
 if env_var is not None:
     pm.cc = env_var
 else:
     pm.cc = "g++"
 pm.fc = None
-pm.meson = True
-pm.mesondir = os.path.join(dstpth)
-
-if sys.platform.lower() == "win32":
-    target += ".exe"
-exe_name = os.path.join(pm.appdir, target)
-
+pm.inplace = True
+pm.makeclean = True
 
 biscayne_cmds = [
     "buildqtg action01_buildqtg.dfn",
@@ -61,16 +65,16 @@ def clean_up():
     pm.finalize()
 
     if os.path.isfile(exe_name):
-        print("Removing " + target)
+        print(f"Removing {target}")
         os.remove(exe_name)
 
-    print("Removing folder " + pth)
-    if os.path.isdir(pth):
+    print(f"Removing folder {pth}")
+    if pth.is_dir():
         shutil.rmtree(pth)
 
     dirs_temp = [dstpth]
     for d in dirs_temp:
-        if os.path.isdir(d):
+        if d.is_dir():
             shutil.rmtree(d)
 
     return
@@ -109,7 +113,7 @@ def run_gridgen(cmd):
 @pytest.mark.regression
 def test_download():
     # Remove the existing target download directory if it exists
-    if os.path.isdir(dstpth):
+    if dstpth.is_dir():
         shutil.rmtree(dstpth)
 
     # download the target
@@ -140,4 +144,4 @@ if __name__ == "__main__":
     test_compile()
     # for cmd in biscayne_cmds:
     #     run_gridgen(cmd)
-    # test_clean_up()
+    test_clean_up()
