+++ autotest/test_mp6.py
