+++ pymake/config.py
@@ -1,6 +1,6 @@
 __author__ = "Joseph D. Hughes"
-__date__ = "March 2, 2022"
-__version__ = "1.2.5"
+__date__ = "December 5, 2022"
+__version__ = "1.2.6"
 __maintainer__ = "Joseph D. Hughes"
 __email__ = "jdhughes@usgs.gov"
 __status__ = "Production"
